pyproject.toml
# SPDX-FileCopyrightText: 2022 UdS AES <https://www.uni-saarland.de/lehrstuhl/frey.html>
#
# SPDX-License-Identifier: CC0-1.0

# https://github.com/psf/black/blob/master/pyproject.toml
# Example configuration for Black

# NOTE: you have to use single-quoted strings in TOML for regular expressions.
# It's the equivalent of r-strings in Python.  Multiline strings are treated as
# verbose regular expressions by Black.  Use [ ] to denote a significant space
# character.

[tool.black]
line-length = 88
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[tool.poetry]
name = "fmi2rdf"
version = "0.1.0"
description = "Extract triples about Functional Mock-up Units (FMUs) in Resource Description Framework (RDF)"
authors = ["Moritz Stüber <moritz.stueber@aut.uni-saarland.de>"]
license = "MIT"
readme = "README.adoc"
repository = "https://github.com/UdSAES/fmi2rdf"
keywords = ["FMI", "FMU", "RDF"]

[tool.poetry.dependencies]
python = "^3.10"
invoke = "^1.6.0"
rdflib = "^6.0.2"
requests = "^2.26.0"
loguru = "^0.5.3"
FMPy = "^0.3.2"
PyYAML = "^6.0"
pydash = "^5.1.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
pytest = "^6.2.5"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
